Resolve "Add "Increase stake" feature for GLDT staking"

+++ backend/canisters/sns_rewards/impl/src/state.rs
@@ -106,7 +106,7 @@ pub struct CanisterInfo {
     pub version: BuildVersion,
     pub commit_hash: String,
     pub memory_used: MemorySize,
-    pub cycles_balance_in_tc: f64,
+    pub cycles_balance_in_tc: u128,
 }
 
 #[derive(CandidType, Deserialize, Serialize, Clone, Copy, Default)]
